Restructured utils.
